Add Text Spectrum app with 3D visualization and React Three Fiber
