+++ README.md
@@ -1,1 +1,15 @@
 # visualizer
+
+This repo contains a React/Vite demo app in `text-spectrum` that turns text into a live 3D color spectrum using React Three Fiber.
+
+## Quick start
+```bash
+cd text-spectrum
+npm ci
+npm run dev
+```
+
+- `npm run build` produces a static site in `text-spectrum/dist`.
+- CI runs lint and build on pushes and PRs.
+
+License: MIT (see `LICENSE`).
