+++ text-spectrum/index.html
@@ -4,7 +4,14 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React + TS</title>
+    <title>Text Spectrum — React Three Fiber</title>
+    <meta name="description" content="Type anything and watch a live 3D color spectrum visualization powered by React Three Fiber." />
+    <meta name="theme-color" content="#070a12" />
+    <meta property="og:title" content="Text Spectrum" />
+    <meta property="og:description" content="Live 3D spectrum of your text in the browser." />
+    <meta property="og:type" content="website" />
+    <meta property="og:image" content="/vite.svg" />
+    <meta name="twitter:card" content="summary_large_image" />
   </head>
   <body>
     <div id="root" class="h-full"></div>
